Improve keeping drawn peaks (for D3) up to date with model.
Switched to listening to the peak model for changes to the peak (instead of manually calling to update peak data after each change), and then marking to load peak data to client side in the next call to render().
Make so when you assign a peak to a nuclide, its color will be changed to the color of other peaks for that nuclide/xray/rctn.