Fix gamma count dialog to use uncert from foreground and background.
Previously when the number of sigma above or below background, only the statistical uncertainty of background would be used; changed it so the uncert of foreground is summed in quadrature with uncertainty of background (and square-rooted) and that is used.  
Added a information icon you can hover over to get a tool tip with the explanation.