Improve users access to RIID results in spec files Created a dedicated dialog in "ShowRiidInstrumentsAna.h/.cpp/.css" to displaying RIID results. Added a notification when file is loaded that alerts the user if there are RIID analysis results available and gives a brief summary of identified nuclides.   The notification also lets the user click a button to see the full analysis results. Added a menu item, "Show RIID Results" to the "View" menu to access the same full RIID result dialog, and also modified the "File Parameters" dialog to use the same display for showing RIID results.
Added a rudimentary check to the "SimpleDialog" class to make sure its contents aren't taller than the users browser, and if so, add some scroll bars.  The method of doing this isn't great (done using the height known by c++), so should maybe be improved for the future.  This was needed incase the list of identified nuclides is longer than can fit on the screen.

Added option to WarningWidget::displayPopupMessageUnsafe(...) function to control how long a notification is displayed for.

Changed the 'InterSpecApp::m_clearSession' to be a slightly more generally useable signal, now named 'InterSpecApp::m_miscSignal' intended primarily for calling back to C++ from buttons in qtip notifications.  Also improved how to handle removing the qtip notification once button is clicked.