Fixes and improvements to Activity/Shielding fit tool.
Fix for possibly using incorrect nuclide Age when nuclides of an element are tied together:
In the Minuit parameters that define how the transport will be done, there is a little bit of a hack that if the age of a nuclide is negative, its value should be interpreted as the index for the nuclide who's age should be used; the problem was the initial nuclide index found was the nuclides index in SourceFitModel, but we actually needed the index to be for the PointSourceShieldingChi2Fcn class, and the ordering can be different (but often times the same).

Activity uncertainty calculation fix:
The activity uncertainty for self-attenuating sources was not be calculated correctly; was using the simple method of just swapping in the uncertainties for the parameters, and taking the result, since this is what can be done for most other fit parameters.  However, since its actually the thickness that is fit for in self-attenuating sources, you need to propagate errors better to account for it being volume that determines activity.  Also the fit nuclide mass fraction uncertainty also needs to be taken into account.

GUI improvements:
- Made the activity and age displayed in center table include uncertainty in same cell
- fixed center source table no always being triggered to update wrt other changes
- Fixed "Fit Mass Fractions" checkbox not always being shown/hidden correctly.
- Added a link to download a text file summary of the calculation log.
- maybe some other small improvements.