Add Electron build handle "deep-links" and become single instance. Changed the Electron build to be a single-instance app; that is, if InterSpec is already open and the user tries to open a spectrum file with it by double clicking on spectrum file, or dragging the file to the EXE, then that spectrum file will be opened in the original instance; this is how macOS, iOS, and Android already handle things. Added support for opening "deep links" on Electron build (tested on Windows only).  That is links with protocol "interspec://" that the user clicks on or scans a QR code, get opened by InterSpec.
Fixed a few small compile issues for on windows, and added open_file_in_session(...) to also accept app-urls.

There is still a little more work to do on having the app be single-instance; like it would be nice to allow multiple windows to be open (but all as part of the same process).