Add cascade-decay sums to Ref Photopeak lines.
Added optison to show cascade decay sum reference lines on the chart.
One note is that x-rays are not included in these cascades (which maybe they shouldnt technically be anyway), only gammas.
But this requires cascade decay data in sandia.decay.xml, which is not in the version nominally distributed with InterSpec, but is avaiable from the SandiaDecay repository (its the ~30 MB xml file).
Also removed scaling of ref. lines based on peak FWHM, and a few other ref. lines. improvements/corrections.