Fix app-menus to be more robust.
Previously, if using websockets transport, we couldnt use JS to display desktop app-menus, leading to a slight, but perceptable delay after clicking on a menu button; the non-web-socket implementation was also not-robust, and had the occasional glitch.
These changes make showing app-menus from JS work for web-sockets, and hopefully is generally a bit more robust, and without glitches.
However, the implementation is still not the greatest and relies on some hacks, like poping up the menus out of visible range at application load.
Getting things to work took far, far longer than I would like, due to dancing around the WPopupMenu behaviour - really, it would have been far better to implement a menu system from scratch (or using a JS framework), and this is just what should probably be done if any major changes are to be done in the future.