Improve number of sig figs displayed on Decay Calc
Changed to tracking the user input activity string (and not just double value), so this way the number of significant digits that should be displayed can be better guessed.
However, this change caused activity to now be tracked by double value, and string, which should probably be cleared up at some point (e.g., remove double value, or make getters/setters for dealing getting/setting the activity.).
Also, right now we can control number of places after decimal when printing out to activity, not the number of sig figs., which should probably be fixed seperately.