Get time chart energy range ratio working.
Finished initial implementation of allowing user to select an energy range to sum events over for the time chart, and then a second energy range to normalize the first one too (usually the energies above the first one).
Seems to work well, but is likely small tweeks will be needed.