Add suggested nuclides to Reference Photopeak tool
Added a column to the Reference Photopeak tool that has "suggested" nuclides in it the user can click.
Right now the suggestions are just from RIID results in the file, and recent nuclides the user has displayed reference lines for.
However, in the future we may include "associated" nuclides (e.g., if you're seeing nuclide X, you should also check for nuclide Y).