Add in parsing of additonal nuc info.
Added in parsing of more_nuclide_info.xml (which is not yet included in the git repo, pending further work), as well checking with this, and enumerating nuclides, to the user, that are enumerated as associated in this file.
Added in stub for creating a "more info window", but creating its content is not yet implemented.