Fix issue using annih. gamma with peaks.
Fix user reported bug that when a peak is assigned as the 511 keV of Na22 (and also applicable to other nuclides), the "Photopeak" area would be empty in some places, and it didnt look assigned.  The issue was the `PeakDef::hasSourceGammaAssigned` function required a transition assigned for nuclides, and wasnt allowing for annih. gammas from nuclide decays.  This also resulted in a few places throughout the code for the peak to not be counted as having a source assigned to it.