Patch over issue of sub-menus possibly being inside parent menus.
When display screen wasnt tall-enough to display all app-menu items, without scroll bars, then sub-menus would be placed inside the parent menu, and just be a total mess.  So a work-around is to just disable vertical scroll-bars in app menus; this is mostly okay because the least important items are at the bottom, and it takes a pretty stout window to be too small to display all the menu items.