Start upgrading reference line data creation.
The logic for showing reference line information, on the chart and in the table is a tangled mess that makes upgrade/improvements a challenge.  This is the start of an effort to refactor things so chart lines backing data structs can be the same as the info table and CSV displays; it also will try to clearly seperate the GUI from the definition and generation of the backing data.

This commit defines some structs to hold the information, and code to fill them out, but this hasnt been tested against old implementation, or actually put to use anywhere.