When suggesting nuclides for peak, let CharacteristicGammas.txt potentially play a little more hefty role.
Previously CharacteristicGammas.txt would increase the weight of a nuclide suggestion, based on all identified peaks, no matter if the peak of interest matched what was in the file or not.
Changed it now so the specific peak of interest is also now checked against the files contents, and if it matches, it gets an extra boost, and if it wasn't previously a candidate isotope, it will be made one.
Weights used are still fairly arbitrary, and I'm sure could be optimized.

Also made it so CharacteristicGammas.txt is only read from file once now, and is stored in memory for later use.