Some work on code quality backing nuclide ID.
Improved the quality of code that suggests peak nuclides, based on nuclide photo peaks lines; this is the code that uses the PhotoPeak.lis file.
PhotoPeak.lis is now read in more efficiently, and only read in once, and then stored in memory.
Previously reactions were effectively not being considered, as well as S.E. and D.E. peaks; now use this.