Convert localhost and web-server targets of app to use same starting of server as the other app target types.
Now use boost::program_options to parse localhost/web-server command line options; also all the setup is now the same as for most of the other apps (except for iOS).  Not sure much was gained, other than some DRY improvements, but this had been a TODO item for some time.