Update background lines to be computed from nuclides.
Previously, "background" reference lines were simply a flat list of ~100 lines.
However, as reported by a user, they were incorrect; normalization was for a decay parent, and not relative to all the nuclide decay chains.
To fix this, now aging and transporting U238, Ra226, Th232, U235, and K40 through a 1m sphere of soil.
Not quite done yet - need to add in escape peaks, and a few other things; maybe also limit the number of lines a bit.