Hopefully finish up initial undo/redo implementation.
Added "Edit" app-level menu with undo and redo menu items; also removed save spectrum as PNG/SVG there as well.  Tested this works for web and macOS, but not electron or wxWidgets.
Added limiting how many undo/redo steps are in memory.  This number can be limited using the app JSON - did not test, or even try to compile for wxWidgets or Electron builds.
Added a `BlockGuiUndoRedo` item to disable undo/redo menu items, and execution of undo/redo steps, during the lifetime of the object; this is to use while tools that do not have undo/redo implemented are open; not perfect, or even great, but out of time for implementing undo/redo.
Fix a few little warnings.