Work around bizarre bug in WebKit.
Sometimes when going to "Make Detector Response" tool, particularly the second time in a session, the Safari or Chrome rendering process would crash, after allocating lots of memory.  This works around this, but its almost unbelievable, so there must be something else going on I don't understand, or perhaps this is just a patch for the one reproducible test case I was using.