Work towards a spectrum QR code that will generate an email.
Now can create a QR code for spectra that contains a 'mailto:' link, but the body of the email may either be a valid URI, or a string of printable text you have to copy/paste in - this is mostly due to the extra layer of URL encoding necessary.  It looks like switching to z85 encoding, instead of ascii85, will save space, so will try this next.

Also fixed up the new export dialog for QR codes.