Fix listen address processing
Whatever skin it's wearing InterSpec is a web app. It smells wrong to me
to change the function signature of start_server and friends depending on
whether BUILD_FOR_WEB_DEPLOYMENT or BUILD_AS_LOCAL_SERVER are defined, when
all the same data files are being used and all javascript and HTML is being
served no matter what display engine is being used.

This diff does a few things:
- makes the function arguments and CLI parameters consistent whether or
  not BUILD_FOR_WEB_DEPLOYMENT is defined. The http-address parameter
  always exists now.
- sets a sensible default value of "127.0.0.1". Unless you know what
  you're doing and why, you probably intend to access InterSpec only
  from the local machine. And unless you specifically give a non-local
  bind address, InterSpec will only listen on localhost.
- Makes `--http-address` optional, and uses the localhost default if not
  otherwise specified
- updates help text to reflect these changes

Tested with
- B_F_W_D/B_A_L_S
- localhost/remote client
- 127.0.0.1, 0.0.0.0, my real IP, a second IP on this machine, the
  hostname of this machine, a hostname that doesn't exist, an IP
  that doesn't exist on this machine, an invalid IP