Less magic, more clarity in http-address handling
Make it more obvious whether this is a local-only or web-deployable build.
A local only build will identify itself in `--help`. A web deployable build
will identify itself differently in `--help`, and will print a message on
stderr when it starts - no matter what address it listens on.

For people who do want to pass the unnecessary `--http-address` argument to
a localhost build, fail immediately if a non-local listen address is given.