Fix up spectrum file query tool to be a little more reasonable.
Made start time, as well as real and live times, be the start time of the IOI spectrum, if it can be identified (instead of just first measurement in file, which could be calibration - or sums of all measurements in file).
Made the CSV download be more friendly to use in Excel.
Fix JS error introduced with Radiacode support.