Add option to right-click on peak and fix it to DRF FWHM.
Now when you right-click on a peak you can select "Use DRF FWHM" to set, and fix, the peaks FWHM to the DRFs value, and then refit the peak.  If no FWHM for the peak is defined, then the `MakeFwhmForDrf` tool will be created to allow making one.
The undo/redo around this probably needs a little more testing, as well as the actual functionality as well.