Start adding peak-skew options.
Added distributions for Bortel, Crystal Ball, Double Sided Crystal Ball, GaussExp, and ExpGaussExp - but all are totally untested (I tested them in Python, so the conversion to c++ is what isn't tested).
Changed a lot of places that need these changes, however, I'm sure there are a ton more.
Have not implemented fitting any of the skew parameters, although did start roughing in the framework.
Code currently compiles, but it will probably explode if you try to use it.