Remove Landau Skew
The Landau skew distribution was poorly setup, and probably as poorly motivate.
There was no support for drawing the distribution to the D3.js based spectrum display (default for the last ~4 years).
It was very questionable how valid the motivation for the distribution was in the first place.
Given the poor setup of the distribution, and requiring 3 parameters, for a single side, it is probably better to just remove it.