Added skew to the "Isotopics by nuclides" tool, and also finding coverage limits for the skew distribution.
Right now the "Isotopics by nuclides" will fit for the various skews, but it doesn't report fit values, or let you to edit them, etc.; there will probably need to be some more work done for that.
The coverage limits for the distributions was motivated to allow drawing better ranges of the peaks in JS, but should be useful a few other place.