Remove option to use native Electron menus.
This option hasn't been used for a number of Electron versions, since the menu mechanism changed.
Just using the HTML menus everywhere besides macOS seems to work/look well, so I don't think it would be worthwhile, or even an improvement to try and get the native Electron menus working again, especially since we'll probably just want to keep the menus in sync with the wxWidgets version of the app.
This commit removed all code in `USE_ELECTRON_NATIVE_MENU` and `USING_ELECTRON_NATIVE_MENU`.
Has not been tested with Electron build yet.