Add log-y option for time chart.
Option only applies to gammas, and to enable you have to click on the three dots menu in the upper right-hand corner of the time-chart, and then go to the "Opt" tab, and click "Log Y Scale".
Only tested with a couple spectra; may need a little more adjusting/fine-tuning.

Closes issue #26.