Add server-side created upload dialog for file drops.
When large files are dropped on the app, there could be a significant gap between the client-side created upload progress closes, and the actual upload being finished at the server, leading to a gap of time that it wasn't clear to the user what was going on, before the parsing status dialog was created.
So now added a dialog triggered by data transfer from the file uploaded - so at least there is something to indicate to the user that things are happening.
Added a timeout of 30 seconds on this dialog (upload request gets cancelled, bad connection, etc).
Also changes to use consistent path to d3.vs.js everywhere, to avoid duplicate network transfers (probably only applicable to debug localhost builds).