Make so if you save or load state from DB, and then subsequently change foreground, and then back to the DB foreground, it will be recognized as working on the state from the DB.
Before, if you save/load a state to/from the DB, and then change foreground, and then back to the DB foreground, you could no longer save to the same state in the DB.
This corrects that, by storing the DB index in SpecMeas, organized by the displayed sample numbers.

Made a few other small changes to wording.