For add_ref_lines.xml, allow specifying reference age for nuclide mixtures.
Now you can specify the age at which the relative activities are effective, and then changing the age of the mixture from this will change the relative activities of the parent nuclides.
Previously I think the activity fractions of the parent nuclides would stay constant, no matter the age the user entered, which probably wasnt the correct thing to do.