Sketch out implementation of batch activity/shielding fitting.
Start implementation of being able to use fit-peaks and shielding/src model from an exemplar, to refit these quantities for different, but similar, files, from the command line.
Right now more of a proof of concept stage - I'm sure a ton of edge cases arent handled, the output is just partially dumped to stdout (no formatting, no reporting, etc), and a ton of things arent complete, or implemented, or correct.