On nuclide search tab, add option to assign all peaks near reference photopeak lines to current selected nuclide.
Before you could assign the peak(s) you used to search for nuclides with, to the currently selected nuclide, but this work extends it to any peaks near that nuclides gammas/x-rays.
Users can still filter results before accepting.
To do this the "Nuc. from Ref." code was re-used, so there were also some improvements to that dialog as well.
Probably needs a little more testing and tweaking.