For file uploads, get rid of `xhr.ontimeout = ...`, which was causing large files to not finish uploading.
`xhr.ontimeout` was being used to remove the upload dialog cover for uploads that stalled more than a minute - this was also causing the upload to timeout after a minute as well, which effectively canceled the upload.  Switched to using a separately defined timer to remove the dialog; now the timer only fires after a minute of no upload progress, and it doesnt cancel the upload.