Fix using non-U nuclides in "Isotopics from peaks" tool when a non SandiaDecay data source is selected.
Fix bug where original data source specification was being used to get nuclide information, when should have been using one that had been corrected for the case where the nuclide wasnt a U nuclide.

Also added in a check to make sure the U-nuclide was in the specified dataset, and if not, fall-back to SandiaDecay.