Add reading QR-codes from HEIC, maybe fix interaction with database.
Added recognizing HEIC files as image files.
For non-JPEG/GIF/PNG files, changed to drawing the uploaded image, at the max of either the images natural size, or displayed size, so if the QR code is only a small portion of the picture, but has enough pixels, it will more likely be found.
Fixed writing to the database in one spot, to be in the wApp main thread, instead of a worker thread.  Not sure if this was an actual bug, but I did encounter a crash, seemingly in this spot (although the crash could be from something else, for sure).