Fix PeakEditor undo/redo support.
The peak editor had a few issues with undo/redo.
If continuum values were changed, then the PeakContinuum object in memory was changed, instead of creating a new one with modified values and using it; so when undo was done, the new values would still be used.  Hopefully fixed this by creating a new copy of the PeakContinuum when its values have been changed.
Also, closing the Peak Editor, when it was opened via undo, would kinda break the undo cycle of things.