Start internationalizing application.
So far, changed the GUI strings in InterSpec.cpp, to use the Wt mechanism to read them from XML files according to the users current language (currently InterSpec.xml, and InterSpec_fr.xml - a LLM translation of the English to French).
Will need to do this same for all the other GUI source files as well - it will be a slog.

Added a "Language" sub menu to the "Help" menu that lets the user select the language, although this menu item will need some work.

Currently the QTip-based tooltips wont dynamically adjust when the user changes the language, but all the other app text that has been converted so far seems to.