Explicitly pass `isHPGe` throughout the app, a bit more.
A number of peak fitting routines implicitly depend if a HPGe detector, or not.  This was previously determined, essentially using the number of channels, instead of possibly better ways.  So this commit some-what improves how the detector resolution is determined, and more explicitly passes it to the various functions.

This is likely a stepping-stone to bigger improvements.