Speed up peak fitting.
For one HPGe spectrum, these improvements decrease the fit time to a little less than a third of before (for release build).
The main thing was getting rid of channels that need evaluating from a `set<size_t>`, to just using ranges (this cut time in half).
Then re-using the same `vector<peakDef>` in each call to  `PeakFitChi2Fcn::chi2(...)` (took another 50% off).
Also slightly optimized call to `erf`.
There is still more to do, but this was the big items.