Added app updating DB with active use time every minute or so.
Before, if app ended unkindly (e.g., crash or killed - like during development), the active use time wouldnt be recorded to the database.  Now this will be recorded every minute, or so.
Also changed it so current values for this, and number of files opened, will be updated from the database, before being incremented, incase multiple sessions are being run simultaneously.