Remove use of moment.js.
moment.js was only used for client-side date/time validation of Spectrum File Query Tool, and it was maybe not even in a working state - so just changed to use server-side date/time validation, which is a bit more consistent with rest of app anyway.  The outdated version of moment.js that was being used was also erroneously triggering some vulnerability scanners.

Also made it so macOS will highlight file when shown in Finder from the Spectrum File Query Tool.