Added command-line batch executable main, and modified CMake build system to for macOS.
It looks like we just need a separate executable for batch analysis (at least for Wx widgets and macOS - tbd for Electron), vs the GUI tool, so added a command-line `main(argv,arc)` interface for batch analysis.

Modified the macOS CMake stuff to to build InterSpecLib as a framework, which gets copied into the app bundle, and both the GUI and batch executables link to the same framework (may eventually link the the qlpreview executable against this same framework).

Windows builds likely wont work.