Some work looking at issue storing preferences in the database.
In debug dev build at least, something really strange is going on with the user preferences, and saving them to the database, and such.  This commit cleans up the preference hygene a little bit, but still experiencing the issue that fter setting preferences, the in-memory values seem to be lost (and actually there are none in-memory) - a total mystery, atm.