Fix crash when exporting peak CSV on wxWidgets build.
`Wt::WServer::instance()` will return nullptr if called from a different
runtime, which was the case for the wxWidgets build in
InterSpecWebFrame.cpp, when the app was built with batch mode, so that
LibInterSpec is shared (and static runtime used to compile, on Windows).
So now get this instance from `InterSpecServer` located within
LibInterSpec.
Tried to improve the 2compile warnings caused by boost::bind, a litte.