Add background spectrum peaks, and background-subtracted peaks to exported peak CSV file.
Now, when you click on the "CSV" link to export peak CSVs, if you have a background peak file loaded, and have fit peaks in it, the exported CSV will contain these background peaks, as well as the foreground peaks, with the background subtracted.  The background subtracted peaks are exactly same as foreground peaks, they just have amplitude decreased by the live-time normalized background peaks within 1-sigma of them, and the amplitude uncertainty is propagated.
Also, for Act/Shield fit, added using the background subtract option, and not just whether peaks were supplied or not.