Get Android app behaving decently with soft-keyboards being shown.
Finally figured out to use `webview.setTranslationY(...)` to move the WebView around when the keyboard is shown; also added in using JavaScript to get the position of the element with focus, so we know how much to move things around.
This was a real battle to figure out and seemingly get working.
Currently the code needs a lot of cleanup, but commiting here to to have a working point.