Add chi2/dof and enrichment text to Rel. Eff. chart
On the Rel. Eff. chart added a chi2/dof (and its p-value) to the chart.  If it s a U-enrich, or Pu-enrich problem, or a ratio of just two nuclides, this relevant info will also appear on the chart.
The chi2 and p-values only use stat uncert, and not the "add. uncert", so if you use additional uncertainty, these values will be terrible.