Fix batch analysis with background subtraction requested when exemplar didnt do background subtraction.
Before, if exemplar didnt have background subtraction enabled, then even if user did request that during batch analysis, it wouldnt be done.
Also, if a spectrum file had a single spectrum, marked as a background, then it wouldnt be used for fitting peaks; now if single spectrum, and the foreground sample is requested, will return that single sample.

Added warnings if no background peaks are actually used during act/shield fit, but it was requested.