Start fixing user options being lost in current session, whenever a refresh or reread from the database is triggered.
It turns out that the object pointed by `Wt::Dbo::ptr` can completely change (i.e., the same smart pointer now points at a different object in memory) when a refresh or reread of the object occurs from the database.  This means you lose any information in the struct pointed to by `Wt::Dbo::ptr`, that isnt directly from the database. Previously, we were caching settings in a std::map, to avoid having to create a database transaction, but these would be lost, because the object pointed to by `InterSpec::m_user` would be changed on us, and then things would default to using the default settings, not what the user selected, or even what was in the database.

This current commit changes this, so settings are kept consistent.
However, the callbacks stored in `InterSpecUser::m_onBoolChangeSignals` still need fixing up.

Also, some performance regressions are observed and should maybe be improved.
For the first time InterSpec is run for a user, setting the initial default preferences takes 60 ms (on a M1 macbook pro - probably due to doing many DB transactions - once per preference), and then just retrieving the preferences after that takes 2 or 3 ms.
If we add-back in caching the settings (with the `InterSpec` class doing the caching), both these times can likely be improved a lot.