Move getting/setting user preferences to a dedicated class, `UserPreferences`.
This separation improves the separation of concerns a bit.
In addition to cleaning up the interface and fixing the callback issues that were still present, it cuts the initial initialization of user preferences down from ~60 ms, to about 1 millisecond; and for users with options already in the database it slightly increases things.

I think this was the last issue that needs to be fixed for the final 1.0.13 release.