Hopefully improve consistency of storing states, and auto-saved spectra in the DB.
Make it so the DB is updated when you switch foregrounds, or user states (before this was only done when spectrum changed through `SpecMeasManager::displayFile(...)`).
The end-of-session temp state should be updated and treated more consistently now.
Add a "dangling" state to user-saved states, that captures unsaved work - there is only one of these per user-saved state.
Re-worked a good amount of the code dealing with DB, with the goal of making it more consistent with this, now a little better thought-out, model.
Removed the "write protected" DB stuff - since it was kinda meaningless - now if something belongs to a tagged state (e.g., has a parent state), then it does not get changed.
Resisted changing the DB schema any, since its so close to a release, but after next release a few improvements should be made.

Currently seems to be working, but hasnt been rigorously tested yet - and their are still a few items (like making sure there are no orphaned files in DB) to be cleaned up.