Fix plotting of double-sided crystal ball distribution not plotting full x-range of peak, for large skew.
Finding the coverage-limits of DSCB distribution may fail for really extended skews (i.e., low power).  This commit extends possible range to search for limits, but it still may fail, in which case the full ROI will now be used (as apposed to 25 sigma previously).
This was only a plotting issue - the full range of the ROI was being used to perform the fit - so no previous peak areas/fits should change, they will now just be displayed properly.

Note: the lower coverage limit of the DSCB function is found analytically, but upper limit is found numerically through bisection - finding the upper limit analytically should be implemented at some point.