Initial work to fix fitting mass-fractions for nuclides of multiple elements.
Tried to refactor most of the fitting code to explicitly group nuclides, who are having mass fraction fit, by elements.
Also changed so individual nuclides can be selected to be fit for or not - the logic for this has been plumbed through the calculation code (although not really tested), and checkboxes added to GUI, but still needs some additional GUI logic to support making a `SourceCheckbox` for the "other" non-src fraction.  The GUI also needs to be cleaned up to look a bit better with these changes.

Code currently compiles, and kinda seems to run (subject to above short-comings), but this was a rather substantial refactor/change, so I am sure there are bugs and problems.