Add Physical Model to to manual Rel Eff calculations.
To the Manual Rel Eff calculations added the FRAM Physical Model, however to get to produce an answer, had to set the Modified Hoerl function to a constant 1.0 - letting the fitter fit these parameters causes the fit to fail.  Not totally sure why yet.  Also, for the "Auto" Rel Eff fitter, setting the same paramters to zero causes it to produce a reasonable result as well.

Did not add physical model to the manual GUI yet, although this doesnt seem too horrible to do.