Get calcs with Modified Hoerl function working, and do work towards adding Physical Model to "Isotopics from peaks" tool.
Added better estimation of initial rel. activities, and a few other tweaks to get fitting with the Modified Hoerl function actually working (at least for a couple examples).
Get most of the way to adding Physical Model to the manual GUI. Created a specialized shielding widget, and mostly plumbed thigns in - but still not yet totally working, and will need tweaks.