Update Minuit2 library to version corresponding to about ROOT 6.34.02.
This update includes the latest version of the Minuit2 library, sourced from https://github.com/root-project/root.
All rights and copyright for the included library code belong to the original authors and the ROOT project.

Other than running the provided CMake utility to copy necessary ROOT files into the Minuit2 source directory, no other modifications were made to the code.
The previous version of Minuit2 used in InterSpec was ~12 years old, and had been manually adapted to build with CMake - however, this update was a drop-in replacement that required no changes to build things, and seemingly works - thats some amazing stability for Minuit2.