Made RelActCalcAuto use a single parameter block, and make struct to represent GUI state.
Going from a seperate paramater block for each paramater, to a single one greatly reduced the number of iterations necassary by Ceres, and should lead to better answers.
Put RelActAuto GUI state into a struct; this means we can use the same XML exported from the GUI, or placed into N42 files to populate this struct for calculations from the code, or unit tests or whatever.