Improve when to use background included with a foreground (hopefully!).
Made it so if you have a background loaded, but you load a new foreground whose file also has a labeled background spectrum, the new background will be used if its start time is closer to foreground than previous background start time (and the previous criteria are all satisfied).
Also added message to user to indicate which action was taken, if it was un-ambigious.