Potentially fix some issues loading app states.
This was motivated by a user report of app getting into a state where the app became unusable, due to loading an invalid state - however, I dont know if this work will actually fix the issue.  The issue was happening in JS, however, I think the JS error was related to a c++ exception causing not everything to be loaded.  I was able to
Have not fixed all potential issues.

Made so InterSpec::loadState c++ has more granular try/catchs, and tried re-arranging order of operations a little so changes to the GUI isnt as intertwined with the DB code, although this is still the case to a good extent.
Fixed issue of loading Act/Shield state that could cause inconsistencies, due to checks being applied to GUI state before it was fully deserialized.
Also added some JS try/catch around the spectrum calls to JS, since this is where the error showed up, but adding these protections are unlikely to be very useful.