Add search by energy for un-aged nuclides, and fix long-standing bug that decay x-rays werent properly included when searching for nuclides.
On "Nuclide Search" tab, added allowing to search for nuclides, based on energies of only the parent nuclide, not of any progeny.
While working on this, also noticed an issue that decay x-rays were not included as candidates for satisfying the search energy requirements.  This bug looks to be there even on the main branch of InterSpec, and not just the revamped Energy search range.