Add column to peak CSV to indicate if peak is Gaussian, or data defined.
In addition to writing this info, reading in a CSV will also pick up this info; if a linear continuum, then the continuum will be re-estimated based on foreground spectrum.
Also added a button to the Peak Editor to estimate the linear continuum values for Data Defined peaks.
These updated CSVs seem to be compatible with older versions of InterSpec.