Integrate differentiating wrt age of nuclide, into the auto-diff mechanism - so now we can use *just* the `ceres::Jet<>` auto-diff mechanism.
Now numerically differentiate the gamma amplitudes wrt nuclide age, then pipe the results into the Jet being used for automatic differentiation.
This approach is much cheaper, and nicer, than using a numeric diff of the evaluation function, as is a relatively very cheap computation, and I think it turns out a lot cheaper than trying to combine the results of numeric and auto differentiation.