Bring out some peak-fitting options to the user when right-clicking on a peak to refit, and in the Peak Editor.
This is part of moving to PeakFitLM as the default (and maybe only) peak fitting routine.
And just generally cleaning things up to make more sense to users (still a way to go here though).

For single-peak ROIs, made it so the peak is always refit with the default PeakFitLM options - e.g., just find the best fit with no real constraints.  I toyed around with allowing a "refine" option, but it just doesnt seem useful.

For multi-peak ROIs, added some options, and hopefully communicated to users a litle what they mean, to fit peaks with the default linear relationship between FWHMs, and an option to let the FWHMs go independent.  In the Peak Editor, users can also choose just a "refine" peak fit, or a combination of options - via a drop-down menu from the "Refit" button.

Will likely need some more work and adjusting as things are used.