Start adding skeleton for batch upload tool.
Change upload method for files drag-n-dropped onto app - now uploads as Form data, which lets Wt take care of spooling the files, instead of us.
Also changed the upload JS around a little; there is more to be done here.
File uploading from drops can be improved further; notes made in JS and C++ for possible future work for this.

Added a fourth upload URL for "batch" uploads; currently set to trigger for more than two files - although this needs to be changed, to be multiple files, that arent set to be foreground/back/second.

Added in skeleton of code to call when batch uploads are initiated - this will eventually turn into the batch widget.